+++ spec/uploadcare/entity/addons_spec.rb
@@ -2,7 +2,6 @@
 
 require 'spec_helper'
 
-# rubocop:disable Metrics/ModuleLength
 module Uploadcare
   module Entity
     RSpec.describe Addons do
@@ -121,4 +120,3 @@ module Uploadcare
     end
   end
 end
-# rubocop:enable Metrics/ModuleLength